Updated package versions
Removed the angular dependencies because they are pulled in by angular-material
and we don't depend on them directly.  This will help resolve some of the
conflicts.  Upgraded to angular-material 1.1.0 which is the latest stable
release as well as angular-formly.  The only breaking change from 7 to 8 does
not affect this project.